feat: extract network utilities
